python/pyproject.toml
[project]
name = "hopsworks"
dynamic = ["version"]
requires-python = ">=3.9,<3.14"
readme = "README.md"
description = "Hopsworks Python SDK to interact with Hopsworks Platform, Feature Store, Model Registry and Model Serving"
keywords = [
    "Hopsworks",
    "Feature Store",
    "hsfs",
    "Spark",
    "Machine Learning",
    "MLOps",
    "DataOps",
    "Model Registry",
    "hsml",
    "Models",
    "ML",
    "Machine Learning Models",
    "TensorFlow",
    "PyTorch",
]
authors = [{ name = "Hopsworks AB", email = "robin@hopsworks.ai" }]
license = { text = "Apache-2.0" }

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Topic :: Utilities",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Developers",
]

dependencies = [
    "pyhumps==1.6.1",
    "requests",
    "furl",
    "boto3",
    "pandas[mysql]<2.4.0",
    "numpy<2",
    "pyjks",
    "mock",
    "avro==1.11.3",
    "PyMySQL[rsa]",
    "tzlocal",
    "fsspec",
    "retrying",
    "hopsworks_aiomysql[sa]==0.2.2",
    "opensearch-py>=1.1.0,<=2.4.2",
    "tqdm",
    "grpcio>=1.49.1,<2.0.0",         # ^1.49.1
    "protobuf>=4.25.4,<5.0.0",       # ^4.25.4
    "packaging",         # ^21.0
]

[project.scripts]
hopsworks-mcp = "hopsworks.mcp.main:main"
hopsworks-cli = "hopsworks_cli.main:main"

[project.optional-dependencies]
python = [
    "pyarrow>=17.0",
    "confluent-kafka<=2.11.1",
    "fastavro>=1.4.11,<=1.11.1",
    "tqdm",
    "hops-deltalake==1.1.2.post2; python_version >= '3.9' and (sys_platform == 'linux' or (sys_platform == 'darwin' and platform_machine == 'arm64'))",
]
sqlalchemy-1 = [
    "pandas<2.2.0; python_version < '3.13'",
    "sqlalchemy<2.0.0",
]
great-expectations = ["great_expectations==0.18.12"]
dev-no-opt = [
    "hopsworks[python]",
    "pytest==7.4.4",
    "pytest-mock==3.12.0",
    "ruff",
    "gitignorefile",
    "pyspark==3.5.5",
    "moto[s3]==5.0.0",
    "pandas>=2.2.0",
    "typeguard==4.2.1",
    "delta-spark==3.3.1"
]
dev-pandas1 = [
    "hopsworks[python]",
    "pytest==7.4.4",
    "pytest-mock==3.12.0",
    "ruff",
    "pyspark==3.5.5",
    "moto[s3]==5.0.0",
    "pandas<=1.5.3; python_version < '3.12'",
    "sqlalchemy<=1.4.48",
    "delta-spark==3.3.1"
]
dev = ["hopsworks[dev-no-opt,great-expectations,polars,python]"]
polars=[
    "polars>=0.20.18,<=0.21.0",
    "pyarrow>=17.0",
]
mcp=[
    "fastmcp>=2.10.5; python_version >= '3.10'",
    "httpx",
    "pydantic>=2.11.7",
]
cli=[
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pyarrow>=17.0",
]

[dependency-groups]
docs = [
    "mkdocs==1.5.3",
    "mkdocs-material==9.5.17",
    "mike==2.0.0",
    "sphinx==7.3.7",
    "keras_autodoc @ git+https://git@github.com/logicalclocks/keras-autodoc",
    "markdown-include==0.8.1",
    "mkdocs-jupyter==0.24.3",
    "markdown==3.6",
    "pymdown-extensions==10.7.1",
    "mkdocs-macros-plugin==1.0.4",
    "polars==0.20.31",
    "mkdocs-minify-plugin>=0.2.0",
]
pre-commit = ["pre-commit"]

[tool.uv]
conflicts = [
    [{ extra = "dev-pandas1" }, { extra = "dev-no-opt" }],
    [{ extra = "dev-pandas1" }, { extra = "dev" }],
    [{ extra = "sqlalchemy-1" }, { extra = "dev-no-opt" }],
    [{ extra = "sqlalchemy-1" }, { extra = "dev" }],
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["tests*"]
include = [
    "../README.md",
    "../LICENSE",
    "hopsworks*",
    "hsfs*",
    "hsml*",
    "hopsworks_common*",
    "hopsworks_cli*",
]

[tool.setuptools.dynamic]
version = { attr = "hopsworks_common.version.__version__" }

[project.urls]
Documentation = "https://docs.hopsworks.ai/latest"
Repository = "https://github.com/logicalclocks/hopsworks-api"
Homepage = "https://www.hopsworks.ai"
Community = "https://community.hopsworks.ai"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "java",
    "transformation_test_helper.py"  # File excluded from ruff formatting to maintain formatting required for tests
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.9+ syntax.
target-version = "py39"

[tool.ruff.lint]
# 1. Enable flake8-bugbear (`B`) rules, in addition to the defaults.
select = ["E4", "E7", "E9", "F", "B", "I", "W"] #, "ANN"]
ignore = [
    "B905",   # zip has no strict kwarg until Python 3.10
    "ANN101", # Missing type annotation for self in method
    "ANN102", # Missing type annotation for cls in classmethod
    "ANN003", # Missing type annotation for **kwarg in function
    "ANN002", # Missing type annotation for *args in function
    "ANN401", # Allow Any in type annotations
    "W505",   # Doc line too long
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.isort]
lines-after-imports = 2
known-third-party = ["hopsworks", "hsfs", "hsml", "hopsworks_common"]


[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
